Ultimata autenticazione

+++ Corso.WebApi/Controllers/StudenteController.cs
@@ -1,10 +1,12 @@
 ﻿using AutoMapper;
 using Corso.Service.DTOs.StudenteDTOs;
 using Corso.Service.IServices;
+using Corso.WebApi.Helpers.InjectableHelpers;
 using Corso.WebApi.Models.StudenteModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MiddlewareExceptionHandler.Controllers;
+using MiddlewareExceptionHandler.ExceptionConfiguration;
 using MiddlewareExceptionHandler.ResponseModel;
 using System.Net;
 
@@ -19,6 +21,7 @@ namespace Corso.WebApi.Controllers
     {
         private readonly IStudenteService _studenteService;
         private readonly IMapper _mapper;
+        private readonly TokenHelper _tokenHelper;
 
         /// <summary>
         /// Costruttore del controller Studente.
@@ -26,10 +29,11 @@ namespace Corso.WebApi.Controllers
         /// <param name="studenteService">Servizio per la gestione degli studenti.</param>
         /// <param name="mapper">Mapper per la trasformazione dei modelli.</param>
         /// <param name="logger">Logger per la gestione dei log.</param>
-        public StudenteController(IStudenteService studenteService, IMapper mapper, ILogger<StudenteController> logger) : base(logger)
+        public StudenteController(IStudenteService studenteService, IMapper mapper, TokenHelper tokenHelper, ILogger<StudenteController> logger) : base(logger)
         {
             _studenteService = studenteService;
             _mapper = mapper;
+            _tokenHelper = tokenHelper;
         }
 
         /// <summary>
@@ -40,7 +44,7 @@ namespace Corso.WebApi.Controllers
         /// <response code="400">BadRequest. L'attributo payload sarà null.</response>
         /// <response code="500">Server error. L'attributo payload sarà null.</response>
         [HttpGet]
-        [Authorize(Roles = "Studente")]
+        [Authorize(Roles = "Admin, Studente")]
         [ProducesResponseType(typeof(ApiResponseModel<List<StudenteDTO>>), StatusCodes.Status200OK)]
         public async Task<ActionResult> ElencoStudenti()
         {
@@ -64,39 +68,21 @@ namespace Corso.WebApi.Controllers
         /// <response code="400">BadRequest. L'attributo payload sarà null.</response>
         /// <response code="500">Server error. L'attributo payload sarà null.</response>
         [HttpGet]
-        [Authorize(Roles = "Studente")]
+        [Authorize(Roles = "Admin, Studente")]
         [ProducesResponseType(typeof(ApiResponseModel<StudenteDTO>), StatusCodes.Status200OK)]
         public async Task<ActionResult> RecuperaStudente(int id)
         {
             try
             {
-                StudenteDTO studenteDTO = await _studenteService.GetById(id);
-                return StandardMessageResult(HttpStatusCode.OK, result: studenteDTO);
-            }
-            catch
-            {
-                throw;
-            }
-        }
-
-        /// <summary>
-        /// Crea un nuovo studente.
-        /// </summary>
-        /// <param name="model">Il modello contenente i dati per la creazione di un nuovo studente.</param>
-        /// <returns>L'oggetto <see cref="StudenteDTO"/> creato.</returns>
-        /// <response code="200">Lo studente creato.</response>
-        /// <response code="400">BadRequest. L'attributo payload sarà null.</response>
-        /// <response code="500">Server error. L'attributo payload sarà null.</response>
-        [HttpPost]
-        [Authorize(Roles = "Studente")]
-        [ProducesResponseType(typeof(ApiResponseModel<StudenteDTO>), StatusCodes.Status200OK)]
-        public async Task<ActionResult> CreaStudente([FromBody] CreaStudenteModel model)
-        {
-            try
-            {
-                CreaStudenteDTO creaStudenteDTO = _mapper.Map<CreaStudenteDTO>(model);
-                StudenteDTO studenteDTO = await _studenteService.Create(creaStudenteDTO);
-                return StandardMessageResult(HttpStatusCode.OK, result: studenteDTO);
+                if (_tokenHelper.IsAuthorized(["Admin"], id))
+                {
+                    StudenteDTO studenteDTO = await _studenteService.GetById(id);
+                    return StandardMessageResult(HttpStatusCode.OK, result: studenteDTO);
+                }
+                else
+                {
+                    throw new CustomException("Unauthorized", HttpStatusCode.Unauthorized, "Ops... Qualcosa è andato storto");
+                }
             }
             catch
             {
@@ -113,15 +99,22 @@ namespace Corso.WebApi.Controllers
         /// <response code="400">BadRequest. L'attributo payload sarà null.</response>
         /// <response code="500">Server error. L'attributo payload sarà null.</response>
         [HttpPut]
-        [Authorize(Roles = "Studente")]
+        [Authorize(Roles = "Admin, Studente")]
         [ProducesResponseType(typeof(ApiResponseModel<StudenteDTO>), StatusCodes.Status200OK)]
         public async Task<ActionResult> ModificaStudente([FromBody] ModificaStudenteModel model)
         {
             try
             {
-                ModificaStudenteDTO modificaStudenteDTO = _mapper.Map<ModificaStudenteDTO>(model);
-                StudenteDTO studenteDTO = await _studenteService.Update(modificaStudenteDTO);
-                return StandardMessageResult(HttpStatusCode.OK, result: studenteDTO);
+                if (_tokenHelper.IsAuthorized(["Admin"], model.IDStudente))
+                {
+                    ModificaStudenteDTO modificaStudenteDTO = _mapper.Map<ModificaStudenteDTO>(model);
+                    StudenteDTO studenteDTO = await _studenteService.Update(modificaStudenteDTO);
+                    return StandardMessageResult(HttpStatusCode.OK, result: studenteDTO);
+                }
+                else
+                {
+                    throw new CustomException("Unauthorized", HttpStatusCode.Unauthorized, "Ops... Qualcosa è andato storto");
+                }
             }
             catch
             {
@@ -138,14 +131,21 @@ namespace Corso.WebApi.Controllers
         /// <response code="400">BadRequest. L'attributo payload sarà null.</response>
         /// <response code="500">Server error. L'attributo payload sarà null.</response>
         [HttpDelete]
-        [Authorize(Roles = "Studente")]
+        [Authorize(Roles = "Admin, Studente")]
         [ProducesResponseType(typeof(ApiResponseModel<int>), StatusCodes.Status200OK)]
         public async Task<ActionResult> EliminaStudente(int id)
         {
             try
             {
-                int idDeleted = await _studenteService.Delete(id);
-                return StandardMessageResult(HttpStatusCode.OK, result: idDeleted);
+                if (_tokenHelper.IsAuthorized(["Admin"], id))
+                {
+                    int idDeleted = await _studenteService.Delete(id);
+                    return StandardMessageResult(HttpStatusCode.OK, result: idDeleted);
+                }
+                else
+                {
+                    throw new CustomException("Unauthorized", HttpStatusCode.Unauthorized, "Ops... Qualcosa è andato storto");
+                }
             }
             catch
             {
